Massive clean-up, raised level cap, implemented cooldown system for Capt. Latanica, made Item Tags purchasable, fixed a few small bugs, lowered death penalty's damage penalty.

+++ src/net/sf/odinms/net/channel/handler/SpawnPetHandler.java
@@ -52,7 +52,7 @@ public class SpawnPetHandler extends AbstractMaplePacketHandler {
         }
         MaplePet pet = MaplePet.loadFromDb(player.getInventory(MapleInventoryType.CASH).getItem(slot).getItemId(), slot, player.getInventory(MapleInventoryType.CASH).getItem(slot).getPetId());
         if (pet == null) {
-            MapleInventoryManipulator.removeById(c, MapleInventoryType.CASH, item.getItemId(), (short) item.getQuantity(), false, false);
+            MapleInventoryManipulator.removeById(c, MapleInventoryType.CASH, item.getItemId(), item.getQuantity(), false, false);
             c.getSession().write(MaplePacketCreator.enableActions());
             return;
         }
